+++ src/utils/typescript.ts
@@ -1,25 +1,24 @@
 import * as vscode from 'vscode';
+import * as util from 'util';
 import * as ts from 'typescript';
 import { parent, uriBasename } from './uri';
 
 export function transpile(
   code: string,
-  uri?: vscode.Uri,
+  fileNameOrUri?: string | vscode.Uri,
 ): [string, undefined] | [undefined, ts.Diagnostic[]] {
-  const result = ts.transpileModule(code, {
+  const output = ts.transpileModule(code, {
     compilerOptions: {
       target: ts.ScriptTarget.ES2024,
       module: ts.ModuleKind.None,
     },
-    fileName: uri && uriBasename(uri),
+    fileName: fileNameOrUri && uriBasename(fileNameOrUri),
     reportDiagnostics: true,
   });
 
-  return result.diagnostics?.length
-    ? [undefined, result.diagnostics]
-    : [fixCode(result), undefined];
+  return output.diagnostics?.length ? [undefined, output.diagnostics] : [fix(output), undefined];
 
-  function fixCode({ outputText }: ts.TranspileOutput) {
+  function fix({ outputText }: ts.TranspileOutput) {
     return outputText.replace(
       /Object\.defineProperty\(exports, "__esModule", { value: true }\);\s*/g,
       '',
@@ -27,14 +26,10 @@ export function transpile(
   }
 }
 
-export function transpileOrThrow(code: string, uri?: vscode.Uri): string {
-  const [transpiledCode, diagnostics] = transpile(code, uri);
+export function transpileOrThrow(code: string, fileNameOrUri?: string | vscode.Uri): string {
+  const [transpiledCode, diagnostics] = transpile(code, fileNameOrUri);
   if (diagnostics) {
-    const error = new TranspilationError(
-      'Transpilation failed, check diagnostics.',
-      diagnostics,
-      uri,
-    );
+    const error = new TranspilationError('Code transpilation failed.', diagnostics, fileNameOrUri);
     error.stack = '';
     throw error;
   }
@@ -52,12 +47,23 @@ export function formatDiagnostics(diagnostics: ts.Diagnostic[], uri?: vscode.Uri
 
 export class TranspilationError extends Error {
   public readonly diagnostics: ts.Diagnostic[];
-  public readonly uri?: vscode.Uri;
+  public readonly fileNameOrUri?: string | vscode.Uri;
 
-  constructor(message: string, diagnostics: ts.Diagnostic[], uri?: vscode.Uri) {
+  constructor(message: string, diagnostics: ts.Diagnostic[], fileNameOrUri?: string | vscode.Uri) {
     super(message);
     this.name = 'TranspilationError';
     this.diagnostics = diagnostics;
-    this.uri = uri;
+    this.fileNameOrUri = fileNameOrUri;
+  }
+
+  [util.inspect.custom](_depth: number, options: util.InspectOptionsStylized) {
+    const host: ts.FormatDiagnosticsHost = {
+      getCurrentDirectory: () => process.cwd(),
+      getCanonicalFileName: (fileName) => uriBasename(this.fileNameOrUri || fileName),
+      getNewLine: () => ts.sys.newLine,
+    };
+    return options.colors
+      ? ts.formatDiagnosticsWithColorAndContext(this.diagnostics, host)
+      : ts.formatDiagnostics(this.diagnostics, host);
   }
 }